@@ -1,4 +1,4 @@
-// swift-tools-version:6.0
+// swift-tools-version:6.2
 
 import PackageDescription
 
@@ -14,7 +14,7 @@ extension Target.Dependency {
     static var rfc5322Foundation: Self { .target(name: .rfc5322Foundation) }
     static var rfc1123: Self { .product(name: "RFC_1123", package: "swift-rfc-1123") }
     static var standards: Self { .product(name: "Standards", package: "swift-standards") }
-    static var incits_4_1986: Self { .product(name: "INCITS_4_1986", package: "swift-incits-4-1986") }
+    static var incits_4_1986: Self { .product(name: "INCITS 4 1986", package: "swift-incits-4-1986") }
     static var standardsTestSupport: Self { .product(name: "StandardsTestSupport", package: "swift-standards") }
 }
 
@@ -32,8 +32,8 @@ let package = Package(
     ],
     dependencies: [
         .package(url: "https://github.com/swift-standards/swift-rfc-1123.git", from: "0.0.1"),
-        .package(url: "https://github.com/swift-standards/swift-standards.git", from: "0.1.0"),
-        .package(url: "https://github.com/swift-standards/swift-incits-4-1986.git", from: "0.1.0")
+        .package(url: "https://github.com/swift-standards/swift-standards.git", from: "0.0.1"),
+        .package(url: "https://github.com/swift-standards/swift-incits-4-1986.git", from: "0.0.1")
     ],
     targets: [
         .target(
