@@ -8,21 +8,21 @@
 import RFC_5322
 import Testing
 
-@Suite("RFC_5322.EmailAddress")
-struct RFC_5322_EmailAddress_Tests {
+@Suite
+struct `RFC_5322.EmailAddress Tests` {
 
     // MARK: - Parsing Tests
 
-    @Test("Parse simple email address")
-    func parseSimpleAddress() throws {
+    @Test
+    func `Parse simple email address`() throws {
         let email = try RFC_5322.EmailAddress("user@example.com")
         #expect(email.localPart.description == "user")
         #expect(email.domain.name == "example.com")
         #expect(email.displayName == nil)
     }
 
-    @Test("Parse email with display name")
-    func parseWithDisplayName() throws {
+    @Test
+    func `Parse email with display name`() throws {
         let email = try RFC_5322.EmailAddress("John Doe <john@example.com>")
         #expect(email.displayName == "John Doe")
         #expect(email.localPart.description == "john")
@@ -30,8 +30,8 @@ struct RFC_5322_EmailAddress_Tests {
         #expect(email.addressValue == "john@example.com")
     }
 
-    @Test("Parse email with quoted display name")
-    func parseQuotedDisplayName() throws {
+    @Test
+    func `Parse email with quoted display name`() throws {
         let email = try RFC_5322.EmailAddress("\"Doe, John\" <john@example.com>")
         #expect(email.displayName == "Doe, John")
         #expect(email.addressValue == "john@example.com")
@@ -39,8 +39,8 @@ struct RFC_5322_EmailAddress_Tests {
 
     // MARK: - Creation from Components
 
-    @Test("Create from components without display name")
-    func createFromComponentsWithoutDisplayName() throws {
+    @Test
+    func `Create from components without display name`() throws {
         let email = try RFC_5322.EmailAddress(
             displayName: nil,
             localPart: .init("user"),
@@ -51,8 +51,8 @@ struct RFC_5322_EmailAddress_Tests {
         #expect(email.domain.name == "example.com")
     }
 
-    @Test("Create from components with display name")
-    func createFromComponentsWithDisplayName() throws {
+    @Test
+    func `Create from components with display name`() throws {
         let email = try RFC_5322.EmailAddress(
             displayName: "Jane Smith",
             localPart: .init("jane"),
@@ -65,8 +65,8 @@ struct RFC_5322_EmailAddress_Tests {
 
     // MARK: - RFC 5322 atext Character Validation
 
-    @Test("All RFC 5322 atext special characters are accepted")
-    func allAtextCharsAccepted() throws {
+    @Test
+    func `All RFC 5322 atext special characters are accepted`() throws {
         // RFC 5322 Section 3.2.3 defines atext as:
         // atext = ALPHA / DIGIT / "!" / "#" / "$" / "%" / "&" / "'" / "*" / "+" / "-" / "/" / "=" / "?" / "^" / "_" / "`" / "{" / "|" / "}" / "~"
 
@@ -79,29 +79,29 @@ struct RFC_5322_EmailAddress_Tests {
         }
     }
 
-    @Test("Exclamation mark (!) is accepted in local-part")
-    func exclamationMarkAccepted() throws {
+    @Test
+    func `Exclamation mark (!) is accepted in local-part`() throws {
         let email = try RFC_5322.EmailAddress("user!tag@example.com")
         #expect(email.localPart.description == "user!tag")
         #expect(email.addressValue == "user!tag@example.com")
     }
 
-    @Test("Pipe character (|) is accepted in local-part")
-    func pipeCharacterAccepted() throws {
+    @Test
+    func `Pipe character (|) is accepted in local-part`() throws {
         let email = try RFC_5322.EmailAddress("user|tag@example.com")
         #expect(email.localPart.description == "user|tag")
         #expect(email.addressValue == "user|tag@example.com")
     }
 
-    @Test("All atext characters together in local-part")
-    func allCharsTogetherAccepted() throws {
+    @Test
+    func `All atext characters together in local-part`() throws {
         let allChars = "test!#$%&'*+-/=?^_`{|}~user"
         let email = try RFC_5322.EmailAddress("\(allChars)@example.com")
         #expect(email.localPart.description == allChars)
     }
 
-    @Test("Multiple atext special characters in same address")
-    func multipleSpecialChars() throws {
+    @Test
+    func `Multiple atext special characters in same address`() throws {
         // Test addresses with multiple special characters
         let testAddresses = [
             "user!tag@example.com",
@@ -120,15 +120,15 @@ struct RFC_5322_EmailAddress_Tests {
 
     // MARK: - String Formatting
 
-    @Test("Format email without display name")
-    func formatWithoutDisplayName() throws {
+    @Test
+    func `Format email without display name`() throws {
         let email = try RFC_5322.EmailAddress("user@example.com")
         let formatted = String(email)
         #expect(formatted == "user@example.com")
     }
 
-    @Test("Format email with display name")
-    func formatWithDisplayName() throws {
+    @Test
+    func `Format email with display name`() throws {
         let email = try RFC_5322.EmailAddress(
             displayName: "John Doe",
             localPart: .init("john"),
@@ -138,8 +138,8 @@ struct RFC_5322_EmailAddress_Tests {
         #expect(formatted == "John Doe <john@example.com>")
     }
 
-    @Test("Format email with display name requiring quotes")
-    func formatDisplayNameNeedingQuotes() throws {
+    @Test
+    func `Format email with display name requiring quotes`() throws {
         let email = try RFC_5322.EmailAddress(
             displayName: "Doe, John",
             localPart: .init("john"),
@@ -151,8 +151,8 @@ struct RFC_5322_EmailAddress_Tests {
 
     // MARK: - addressValue Property
 
-    @Test("addressValue returns email without display name")
-    func addressValueProperty() throws {
+    @Test
+    func `addressValue returns email without display name`() throws {
         let email = try RFC_5322.EmailAddress(
             displayName: "John Doe",
             localPart: .init("john"),
@@ -163,36 +163,36 @@ struct RFC_5322_EmailAddress_Tests {
 
     // MARK: - Validation Errors
 
-    @Test("Reject email without @ sign")
-    func rejectMissingAtSign() throws {
+    @Test
+    func `Reject email without @ sign`() throws {
         #expect(throws: RFC_5322.EmailAddress.ValidationError.self) {
             _ = try RFC_5322.EmailAddress("userexample.com")
         }
     }
 
-    @Test("Reject consecutive dots in local-part")
-    func rejectConsecutiveDots() throws {
+    @Test
+    func `Reject consecutive dots in local-part`() throws {
         #expect(throws: RFC_5322.EmailAddress.ValidationError.self) {
             _ = try RFC_5322.EmailAddress("user..name@example.com")
         }
     }
 
-    @Test("Reject leading dot in local-part")
-    func rejectLeadingDot() throws {
+    @Test
+    func `Reject leading dot in local-part`() throws {
         #expect(throws: RFC_5322.EmailAddress.ValidationError.self) {
             _ = try RFC_5322.EmailAddress(".user@example.com")
         }
     }
 
-    @Test("Reject trailing dot in local-part")
-    func rejectTrailingDot() throws {
+    @Test
+    func `Reject trailing dot in local-part`() throws {
         #expect(throws: RFC_5322.EmailAddress.ValidationError.self) {
             _ = try RFC_5322.EmailAddress("user.@example.com")
         }
     }
 
-    @Test("Reject local-part exceeding 64 characters")
-    func rejectTooLongLocalPart() throws {
+    @Test
+    func `Reject local-part exceeding 64 characters`() throws {
         let longLocalPart = String(repeating: "a", count: 65)
         #expect(throws: RFC_5322.EmailAddress.ValidationError.localPartTooLong(65)) {
             _ = try RFC_5322.EmailAddress("\(longLocalPart)@example.com")
