@@ -8,21 +8,21 @@
 import Testing
 @testable import RFC_5322
 
-@Suite("String Conversions")
-struct String_Tests {
+@Suite
+struct `String Tests` {
 
     // MARK: - EmailAddress to String
 
-    @Test("Convert simple email to string")
-    func emailToString() throws {
+    @Test
+    func `Convert simple email to string`() throws {
         let email = try RFC_5322.EmailAddress("user@example.com")
         let string = String(email)
 
         #expect(string == "user@example.com")
     }
 
-    @Test("Convert email with display name to string")
-    func emailWithDisplayNameToString() throws {
+    @Test
+    func `Convert email with display name to string`() throws {
         let email = try RFC_5322.EmailAddress(
             displayName: "John Doe",
             localPart: .init("john"),
@@ -33,8 +33,8 @@ struct String_Tests {
         #expect(string == "John Doe <john@example.com>")
     }
 
-    @Test("Convert email with quoted display name to string")
-    func emailWithQuotedDisplayNameToString() throws {
+    @Test
+    func `Convert email with quoted display name to string`() throws {
         let email = try RFC_5322.EmailAddress(
             displayName: "Doe, John",
             localPart: .init("john"),
@@ -45,8 +45,8 @@ struct String_Tests {
         #expect(string == "\"Doe, John\" <john@example.com>")
     }
 
-    @Test("Display name with special characters gets quoted")
-    func displayNameWithSpecialCharsQuoted() throws {
+    @Test
+    func `Display name with special characters gets quoted`() throws {
         let email = try RFC_5322.EmailAddress(
             displayName: "John@Doe",
             localPart: .init("john"),
@@ -60,8 +60,8 @@ struct String_Tests {
 
     // MARK: - Message to String
 
-    @Test("Convert message to string")
-    func messageToString() throws {
+    @Test
+    func `Convert message to string`() throws {
         let message = RFC_5322.Message(
             from: try RFC_5322.EmailAddress("sender@example.com"),
             to: [try RFC_5322.EmailAddress("recipient@example.com")],
@@ -79,8 +79,8 @@ struct String_Tests {
         #expect(string.contains("Hello"))
     }
 
-    @Test("String conversion matches byte conversion")
-    func stringMatchesByteConversion() throws {
+    @Test
+    func `String conversion matches byte conversion`() throws {
         let message = RFC_5322.Message(
             from: try RFC_5322.EmailAddress("sender@example.com"),
             to: [try RFC_5322.EmailAddress("recipient@example.com")],
@@ -98,8 +98,8 @@ struct String_Tests {
 
     // MARK: - DateTime to String
 
-    @Test("Convert datetime to string")
-    func datetimeToString() throws {
+    @Test
+    func `Convert datetime to string`() throws {
         let dateTime = try RFC_5322.DateTime(
             year: 2024,
             month: 1,
@@ -115,8 +115,8 @@ struct String_Tests {
         #expect(string.contains(":"))
     }
 
-    @Test("DateTime string matches description")
-    func datetimeStringMatchesDescription() {
+    @Test
+    func `DateTime string matches description`() {
         let dateTime = RFC_5322.DateTime(secondsSinceEpoch: 1609459200)
 
         #expect(String(dateTime) == dateTime.description)
@@ -124,16 +124,16 @@ struct String_Tests {
 
     // MARK: - Header to String
 
-    @Test("Convert header to string")
-    func headerToString() {
+    @Test
+    func `Convert header to string`() {
         let header = RFC_5322.Header(name: .contentType, value: "text/plain")
         let string = String(decoding: [UInt8](header), as: UTF8.self)
 
         #expect(string == "Content-Type: text/plain")
     }
 
-    @Test("Header string format")
-    func headerStringFormat() {
+    @Test
+    func `Header string format`() {
         let header = RFC_5322.Header(name: "X-Test", value: "test value")
         let string = String(decoding: [UInt8](header), as: UTF8.self)
 
